+++ glean_parser/kotlin.py
@@ -199,6 +199,9 @@ def output_kotlin(objs, output_dir, options={}):
         'allowed_extra_keys',
         'disabled',
         'include_client_id'
+        'range',
+        'bucket_count',
+        'histogram_type'
     ]
 
     namespace = options.get('namespace', 'GleanMetrics')