+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 Unreleased
 ----------
 
+* `custom_distribution` metric type has been added.
+
 * `labeled_timespan` is no longer an allowed metric type.
 
 1.0.0 (2019-07-29)